Quiet forVersusWhilePerf regressions
PR #6833 changed forVersusWhilePerf.c to be more comparable to the Chapel
version. It changed some int's to int64_t's, but used the wrong printf
formatter and didn't include stdint.h.

Here I include stdint.h and cast the results in printf to long ints. I could
have used the PRId64 formatter, but I never remember how portable that is, and
we don't use it in any other tests, so I opted to just cast in order to quiet
the regression for now.