With comm=ugni, register large arrays with the NIC dynamically.
This is a preliminary implementation of non-heap allocation and dynamic
NIC registration for arrays.  It is used in the comm=ugni configuration
when there is a craype-hugepage* module loaded.  Currently it only
applies to arrays that are at least 2 hugepages in size.  The major
benefits of this feature is that it allows first-touch to achieve good
NUMA locality for such arrays.  It also means that the registered heap
can be much smaller by default, since large arrays are not allocated
from it.  The major downside is increased overhead for both array
allocations and dynamic registration bradcasting.

This feature improves stream and stream-ep performance with comm=ugni
and a registered heap by somewhat more than 2x, bringing it roughly up
to par with comm=gasnet and the MPI substrate, our best-performing
configuration for these tests.

A more detailed breakdown of the work is as follows.

In the runtime comm layers:
Add new chpl_comm_regMem*() functions to work with comm layer registered
memory: threshold size, allocate, post-allocate (register), and free.
Arrange for these to be an optional part of the interface, and implement
them only for comm=ugni.  Note that the implementation is a somewehat
preliminary one.  In particular, the broadcasts of updated memory maps
when arrays are registered and deregistered on one node are done by
means of serial executeOn operations to each of the other nodes to
update their map entries for that nodes' memory regions.  In addition,
the number of arrays that can be dynamically registered is fairly small,
in practice about 15 or so.  Finally, for this to be effective the user
must set HUGETLB_NO_RESERVE=yes in the environment in order for hugepage
locality to follow first-touch.  (This latter is to get around a kernel
issue which we hope will be fixed soon.)

In the runtime memory layers:
Update chpl_mem_array_alloc() and its wide partner to try to allocate
through comm layer if big enough.  If successful, tell the caller to
call back again after initializing, passing back the allocated pointer.
At that point, call the comm layer's post-allocate routine.  (This back
and forth dance is temporary and will be fixed by future changes to how
the "array_alloc" primitive works.  Then, in chpl_mem_array_free(), try
to free through the comm layer and fall back to freeing to the regular
memory layer code iff that fails.

In the module code:
Adapt ChapelBase.chpl:_ddata_allocate() to the changed interface to
chpl_mem_array_alloc(), via the "array_alloc" primitive.  And then in
DefaultRectangular.chpl, adapt the _ddata_allocate() calls to its new
interface.

In the compiler:
Adapt the codegen logic for primitive("array_alloc") to the new
chpl_mem_array_alloc() interface.

While here, in the comm=ugni alignment macros, debugging, and internal
statistics reporting, I made some minor clean-ups and simplications that
arose out this work.

There is a fair amount of future work that could or should follow this:
It may be desirable to allow users to enable/disable this feature by
means of a config const, presumably in ChapelBase.
The array size threshold should be tunable.
The comm=ugni registration update broadcast logic needs to be made
scalable.
It would probably be safe to set HUGETLB_NO_RESERVE automatically so the
user didn't have to.
It might be nice to add a way to set "interleaved" memory binding for
the hugepages, to improve performance for applications that are not
sensitive to sublocale (NUMA) locality but can benefit from the added
bandwidth when all the NUMA memory pathways are used.
The comm layer "post-allocate" function could be called "register"
instead.