Adjust call destuctors to correctly deinit in try blocks
See the new err-return test for an example where the new behavior is
relevant.

Call destructors needs to be aware of error hadling because it has both
control flow impact and also because it has an impact on where
deinitialization should occur. In particular variables local to a try
block should be deinitialized before the catch blocks execute.
Additionally, calls that throw in a try block should deinitialize locals
in that try block if an error occured - but not a local that was not
fully initialized (as in the case when a function returning a record
throws an error).