Add a couple more tests of named expressions in initializers
One verifies that named arguments can be used explicitly in initializers (as
the other tests we have all involve omitted field initialization).  The other
checks the behavior when using a named argument to pass "this" to a function
when we shouldn't - this unfortunately results in an internal error instead of
the appropriate error message.