Add debugGetTheIteratorFn()
debugGetTheIteratorFn() allows us to get conveniently from a ForLoop
id or an iterator type to the corresponding iterator FnSymbol in the
debugger. I included a special adjustment that returns the iterator
corresponding to the _getIterator function, if the latter is
encountered first. Compared to getTheIteratorFn(), the "debug" variety
is safer for use in the debugger because it simply returns NULL where
the production version would trip over one of many INT_ASSERTs.

debugGetTheIteratorFn() is now also invoked from whocalls() to detect
if the FnSymbol is used as the iterator in a ForLoop.

Aside: I placed two key debugGetTheIteratorFn implementations in
lowerIterators.cpp so they can be kept up-to-date with getTheIteratorFn().