Remove -pthread from LLVM link step (#6919)
Remove -pthread from LLVM link step (#6919)

[reviewed by @mppf ]

The -pthread flag is needed when compiling with LLVM. However, when linking, it is not needed. In the last couple of weeks, a new version of clang installed on Macs has started issuing a warning message when the flag is used where it is not needed. This causes the Chapel tests to fail with LLVM on Macs because of the extra output text.

This PR keeps the -pthread flag when compiling, but removes it when linking. This was tested across Mac, Cray, and Linux systems in various configurations that included the different third-party packages.

If we ever do link-time optimization, where LLVM emits a .bc file instead of a .o file from a compilation, it is possible that we will need to add this flag back.