Improvements in dynamic registration with many registered arrays.
We observed problems with isx and isx-hand-optimized timing out with
dynamic registration.  Ultimately there were two problems, one a limit
that was arguably too low and the other an actual bug.  The undervalued
limit was the size of the table of registered memory regions.  It only
had room for 20 entries, including the typical half-dozen or so "fixed"
entries for the static data segment, heap, and so on.  Isx has an array
per core, so on any compute node with more than about 15 cores we would
run out of table entries.  Here, increase the table size to 100.  Using
that many entries will magnify the overhead introduced by searching the
table serially, but we knew we might have to address that overhead in
the future anyway.

The second problem was that once we ran out of table entries and so
dropped back to allocating out of the regular heap, in the regMemFree
routine the comm layer could mistakenly think it was being asked to free
memory it had allocated when in fact that memory was from the heap.  The
old logic looked for any memory region that contained the address being
freed.  If it found one then it assumed the comm layer had allocated the
space.  But this would be the right test only if all array allocations
were done by the comm layer.  If any of them could be done by the memory
layer and thus have been satisfied out of the regular heap, as would be
the case if we ran out of registered memory table entries due to a limit
that was too low, then we would try to deregister the corresponding
space in the middle of the heap, and would have tried to release the
hugepages in the middle of the heap as well.  So, change regMemFree's
test to include checking the base and size of the memory region it
finds, and decide that the space was allocated by the comm layer if and
only if those two exactly match what is being freed.