Remove now-redundant ifi-baseline test
ifi.chpl used to be a future that passed with --baseline.  So ifi.chpl
had a skipif baseline.  ifi-baseline.chpl was a copy of ifi.chpl that
ran with baseline to lock in it passing.

Now that ifi.chpl passes with and without --baseline and is no longer
skipped with --baseline, ifi-baseline.chpl is redundant with it.