Fix valgrind failures by using autoCopyMap::find to test for membership
Directly indexing with operator[] will create an entry with Type* as a
key. If that Type* is later deleted, future accesses of autoCopyMap will
compare against that key with std::less. We define std::less for various
BaseASTs with the def_less_ast macro, which will query the 'id' field
of the given arguments. If an argument was deleted, this will result in
valgrind failures.