Work around kernel bug, adjust test-specific heap sizes.
(This patch set from @ronawho -- thanks!)

We need to set HUGETLB_NO_RESERVE in order to work around a kernel bug
that sets NUMA locality for hugepage space even before it's touched.  Do
so in the Chapel launcher; the system launcher will copy this setting
down to the compute node(s).  We have hope that the kernel bug will be
fixed at some point, at which time we can remove this.  (This would be
desirable because the workaround has the side effect of causing programs
to get a SIGBUS if they over-reserve the hugepages, rather than simply
getting a synchronous allocation failure.)

Also, unset the heap size setting for a number of tests that allocate
really big amounts of space.  These settings are now not needed and
indeed may cause us to fail with the HUGETLB_NO_RESERVE-related SIGBUS
on aprun systems, because they limit how many hugepages are available
for non-heap allocations.