Add user-facing error for exported function with generic arguments
As reported in issue #6945, using a generic argument in an external
function was previously causing an internal error.  This specializes
that internal error for the case that the function is marked `export`.
I have a strong suspicion that other types of exported generic
functions may not be caught by this code path.