Fix using_chapel_module()
In #6946 I changed using_chapel_module() to do a normpath on CHPL_HOME and
CHPL_MODULE_HOME before comparing them. If both were unset, this would
incorrectly result in the function thinking the chapel module was being used,
which resulted in building the runtime for the lowest common denominator on the
whiteboxes.

This updates using_chapel_module() to only do the normpath after checking if
CHPL_HOME and CHPL_MODULE_HOME were actually set.