Implement serialization for RVF and global broadcasting
Prior to this commit, RVF could only bit-copy records across on-statements.
Bit-copying is insufficient for records that contain classes, arrays, or other
heap-allocated memory. This commit implements support for user-defined
serialization across on-statements. This serialization support can be disabled
with the new compiler flag ``--[no-]remote-serialization``.

=============
The Interface
=============

Users can implement two methods that the compiler will use for serialization.
The compiler will only identify these methods if they are implemented on a
record.

1) The ``proc chpl__serialize()`` method accepts no arguments and returns
something that can be auto-destroyed. The compiler does not currently do any
kind of automatic memory management for classes, so the result must be
auto-destroyable in order to avoid leaks.

This data will be bit-copied into an on-statment's arg bundle.

2) The ``proc type chpl__deserialize(data)`` method is a **type method** that
accepts a single argument. This argument is the result of a call to
``chpl__serialize()``. This method must return the same type as its parent
record.

If a ``chpl__serialize`` method is found without a matching
``chpl__deserialize``, the compiler will emit a warning.

=================
Program Execution
=================
If the compiler determines that a variable can be forwarded in the
remoteValueForwarding pass, it will insert a call to ``chpl__serialize`` before
each call to the on-statement. The result will replace the variable in the
on-statements formals.

The compiler will then insert a call to ``chpl__deserialize`` at the beginning
of the on-statement and pass the serialized data as an argument. The result
will be used in place of the forwarded variable throughout the on-statement.

The result of ``chpl__deserialize`` and the serialized data itself will both
be destroyed at the end of the on-statement.

====================
Broadcasting Globals
====================

If the compiler resolves a ``chpl__serialize`` and ``chpl__deserialize``, it
will utilize them when broadcasting global consts. Instead of a call to
``PRIM_PRIVATE_BROADCAST``, the compiler will insert a call to a new function
named ``chpl__broadcastGlobal``. This function is defined in the new internal
module ``ChapelSerializedBroadcast``. ``chpl__broadcastGlobal`` is responsible
for calling ``chpl__serialize`` on the root locale and for moving the result
of ``chpl__deserialize`` into the private global on other locales.

To support this, a new C table named ``chpl_global_serialize_table`` has been
introduced. This table contains pointers to globals. This table is used by
``chpl__broadcastGlobal`` in order to move the deserialized result into
the privatized global. The index-->global mapping is decided at compile-time
in the ``parallel`` pass.

If an autoDestroy for that global's type exists, the compiler will insert a
call to ``chpl__destroyBroadcastedGlobal`` before the original global is
destroyed. This function is responsible for calling chpl__autoDestroy on the
corresponding serialized global on other locales.

In this initial commit, a simple ``coforall loc in Locales do on loc`` is used,
but a tree-spawning pattern would likely perform better.

======
Impact
======

This initial commit implements serialization for DefaultRectangular domains.
Initial tests suggest reduced communication, but more testing is required to
determine actual performance benefits.