Add two tests of missing error messages for initializers
One covers method calls in Phase 1 when the call is within an if expression, the
other covers modification of the contents of a const array field during Phase 2.
Both error messages are generated in ordinary circumstances but are avoided in
these more complicated examples.

Brad ran into these when updating one of the constructor tests to use
initializers.