Added an error message for size mismatches in tuple casts
This resolves issue #6961 in which returning a tuple of the wrong
size resulted in an internal error in some cases and the ignoring
of the extra fields in other cases.  Here, we do an explicit size
check before entering the code in question and issue an error if
there's a mismatch.