Refactor based on feedback
Mike suggested that I was working too hard to specialize
hasGenericFormal() without sufficient motivation, and on reflection, I
agreed, so rather than trying to make it general for a bunch of
nonexistent calling contexts, I just moved the current error into the
routine.  In doing so, I decided to split the error into multiple
lines so that if a function was generic due to a number of arguments,
there error would be a bit more precise.  Updated my test to include
such a case and updated the .good file to match.