Stop forcing pre C++11 ABI for cray-prgenv builds
gcc 5.1 introduced a new libstdc++ ABI to conform with the C++11 standard, but
it was incompatible with the old ABI. When we initially upgraded our gen
compiler to gcc 5.1 we still wanted to be able to use gcc 4 as a target
compiler, so we built with `-D_GLIBCXX_USE_CXX11_ABI=0`.

However, GLIBCXX_USE_CXX11_ABI=0 is now causing link errors for newer versions
of cce, so this just stops forcing the old ABI. The last gcc 4.X is over 2
years old at this point, and a Cray (especially one that has Chapel installed)
will have a newer default.

This effectively reverts #3565, which also has more details on the original
motivation for throwing `-D_GLIBCXX_USE_CXX11_ABI=0`.