Add two tests of initializers that I hadn't handled before
One verifies the behavior when field initialization is surrounded by a sync
statement (this should be okay because sync statements do not introduce
parallelism or loops by themselves).  The other verifies the behavior when
method calls are present in Phase 1 but the method being called has multiple
versions (which in the current implementation would mean we don't know the
symbol at scopeResolve, so I expect we won't recognize it is a method call
at the time we verify initializer rules).