Have test use an explicit binary name
This test core dumped in one configuration last night, and my best
guess as to why is that the test tries to overwrite the executables
that are being used to run it as it runs which could cause problems
(maybe?).  Here, I'm working around that by having the test specify
an explicit executable name and removing that executable between
compiles.