Enable remote serialization for strings
This is easily accomplished by using existing infrastructure for copy
remote string data, but we would prefer to avoid the GET for small
strings that could go directly into the arg bundle.

To accomplish this, I have created an external struct named
``chpl__inPlaceBuffer`` with a fixed-size C array as its only field. This
memory will be used when the string is sufficiently short. This cutoff is
represented by the CHPL_SHORT_STRING_SIZE macro.

The ``chpl__getInPlaceBufferData`` method is used instead of specifying an
extern field due to an initialization issue. If the field is declared to be
a ``c_ptr(uint(8))``, the initializer will attempt to initialize it with a
value of NULL. The backend compiler will throw an error in that case.

Also enables broadcasting of global strings (but not string literals).