use error handling in standard IO module (#6963)
Use the IO module as a major test case for the error handling system, in turn making the module more resilient.

leave syserr intact:
The IO module uses syserr as a representation for error codes returned from the runtime, passing them around internally and out externally. Replacing each function that takes an out syserr with throws and strict error handling would require extensive changes by library users, outside the intended scope of this change.

ioerror():
Instead of replacing each instance of syserr, this change focuses on the function ioerror(), which was used in the library to halt the program on certain system errors. This functionality has been changed to throw a SystemError, a subclass of Error that contains a message about the syserr. Some intentionally failing tests had their output changed because the halt() is done through the error handling system instead of by ioerror() directly.

throws, try, try! throughout:
throws was added to any function that called ioerror(), along with the try annotation so that the IO module itself would compile in strict mode. This naturally extended to any functions that called those throwing functions. However the throwing functions can be safely ignored by users in default mode.

results and outlook:
linux64+flat testing is clean in default mode and fails 1529/5962 tests in strict mode, which was expected given that many tests do not wrap with try. The goal is to eventually replace syserr with the error handling system, but it would require deprecation warnings so that users can have a smooth transition.