chplspell: Skip test/studies/shootout
The "submitted" subdirectory contains the shootout benchmarks as
submitted.  The other directories contain shootout development
projects.

Typos in the former won't be corrected, as they're meant to be an
exact match of the last submitted benchmarks.  Typos in the latter are
uninteresting.  Before being submitted, they'd be placed in
test/release/examples/benchmarks/shootout, which is still included in
chplspell's DEFAULTDIRS.