Fix compflags/diten/savec.chpl test for CHPL_GASNET_SEGMENT=fast
Any use of the extern function 'system' causes seg faults for gasnet-fast.
Change 'system' calls into calls to a function that uses the 'Spawn' module to
do something equivalent. Ran the modified test 100 times with gasnet-fast.

Add a simple .future that only runs for gasnet-fast that calls 'system' and
crashes.