Enlarge registered memory tables and report if/when they overflow.
In comm=ugni we record the registered memory regions in a fixed-size
table.  Here, greatly increase that table's size (from 100 entries to
1000) and add a warning emitted the first time we discover that we want
more entries but don't have any.  The ISx test, when run on compute
nodes with high core counts, was silently running out of such table
entries, dropping back to allocating arrays out of the regular heap, and
then running out of memory.  With this change it should no longer do
this, but if it or any other test overflows the table at least we'll
know that has happened.

Note that expanding the tables arguably makes it even more imperative
that we improve the scalability of the broadcast and search algorithms
used on them.