Change the name of the 'Barrier' module to the plural 'Barriers'
Previously, the 'Barrier' module and the 'Barrier' record shared a name, which
could lead to problems, as discussed in issue #6979. Update the name of the
module to 'Barriers'.

One other idea was to change the name of the record to all lowercase 'barrier',
and change the constructor to an 'init' function, but then the 'barrier'
method looked like a constructor, so would also need to change names.
Updating the module name appears to be a lower impact change.

Update a few tests to use the new module name.