Remove -Winline from GASNET_OPT_CFLAGS
We previously had been removing -Winline from the other GASNet flags, but in
practice it seems to come in GASNET_OPT_CFLAGS, so remove it there too.

This commit leaves the old removal of -Winline from CHPL_GASNET_CFLAGS_ALL in
place in case -Winline can appear there in the future or in some other
configuration.