Add the 'Barrier' module back with a deprecated warning
The 'Barrier' module just issues a 'compilerWarning' saying it is deprecated
and does 'use Barriers;'

Added a deprecated test to make sure the 'compilerWarning' is happening.

Added Barriers.chpl to the list of files to chpldoc.  Barrier.chpl is still
on the list, but now just says it is deprecated and links to Barriers.