Add annotations for this week
Emitted code size and AST size increased, while the binary trees release
version and bradc version, and the reductions timings test improved as a result
of adding Error handling to the IO module.

Also tracked down a slight slow down/code size increase in the last week of
July to be partially due to the presence of the new dsiAssignDomain function
and moving endcounts to task local storage.