Future for distributed bulk transfer on rank change
This test demonstrates that something about the distributed bulk
transfer interface breaks down when rank change is involved.  It
may also have required me to use a subset of the dimensions that
are preserved, though I don't feel confident about that as I
type this.  At a quick glance, it appears to me that there's no
conversion going on in the domain between low- and high-D when
being passed through the doiBulkTransfer() interface.