Fix bug with proper field dependencies
I was accidentally copying over references to the fields explicitly rather than
to the new arguments I had created to represent them.  This changes fixes that
for default initializers and also for the _new function because we were
accidentally doing the same thing there.

After this change, proper field references work.  Out of order field references
still fail in an unclear manner, and references to parent fields aren't
handled yet.  This allows the initializer conversion I made for the last
constructor in the madness suite to work.