Add ForLoop::isLoweredForallLoop
The error handling pass needs to be aware of which ForLoops represent the
standalone/leader part of a forall loop so that it can perform
appropriate transformations. To enable that, ajusts ForLoop to track this
information. The new member field is simply set at ForLoop construction
time appropriately depending on what the forall loop represents.

Additionally, used the new isLoweredForallLoop() in cullOverReferences
and assert that it identifies the same forall loops as the previous
strategy that code was using. That is simply meant as a way to improve
testing of the new isLoweredForallLoop() function.