Add GOTO_BREAK_ERROR_HANDLING as a type of goto
Certain error handling constructs with forall statements need to be able
to exit a forall loop. The error handling gotos are a little bit delicate
because they are used/adjusted in both callDestructors and
lowerIterators. In this case, we need to be able to exit a forall loop
without also destroying the local variables in scope. In that sense, we
need an error-handling related loop exit for forall loops that has more
of the flavor of GOTO_BREAK.

It is possible that GOTO_BREAK would be sufficient in this case but I'm
prefering to leave it as GOTO_BREAK_ERROR_HANDLING as it has some of the
properties of GOTO_BREAK and some of the properties of
GOTO_ERROR_HANDLING.