Adjust errorHandling lowering to solve forall issue
Previously to this commit, the combination of error handling
lowering and call destructors was calling _freeIterator
at the time a forall loop was exited with an error - that is,
within one of the coforall iterations (in the typical case
where a forall is implemented with coforall and for).

Instead, the _freeIterator call (and the auto-destroy calls for
local variables declared at the same scope as the forall loop itself)
needs to happen in a loop cleanup block, not as part of exiting the loop.

The simplest way to do that that I was able to come up with was to modify
the error handling pass to use a pattern more like `break` to exit from
the forall loop, and then to perform the usual destruction after that
happens. And then to handle the error in the parent scope. This commit
makes that change.

Additionally it factors out two helper functions for use in
adding forall handling and for regular error handling. It creates
setOuterErrorAndGotoHandler and also gotoHandler methods.