Update last of the madchap helper constructors to be an initializer
Prior to the update I made to the _new function, this change was encountering
a segfault in copyPropagation.  It turns out that the _new function was not
properly copying arguments, especially in the case of argument dependencies - in
this case it was trying to reference the 'k' argument in the init function
instead of the corresponding 'k' argument it added to itself.  Now that I have
resolved that error, I can commit this change.

This test would benefit from the ability to update const fields during Phase 2.