Add curly braces around code before printMemAllocs
Note that printMemAllocs prints out currently allocated memory,
and this test is meant to exercise that function.

PR #6942 changed the way that _freeIterator calls are added. In
particular, it changed to use DeferStmt to add these calls. That
generally improved correctenss but created an issue with this test in
--baseline configurations.  In that configuration, the _freeIterator
statement was added just after printMemAllocs when before the PR it was
added just before. The reason for this has to do with the way that
DeferStmt works - namely, it adds the defered statement at the end of the
block. Currently, our forall scaffolding (at least in --baseline) doesn't
add an enclosing block, so the block that the DeferStmt applied to was
the block containing printMemAllocs, so the _freeIterator call ran after
it).

While it's conceivable that this issue might impact other tests, it semes
to only occur in --baseline. Additionally, the memory is getting freed,
just not exactly where it was before. Thus I'm viewing it more as a
curiousity and adding a simple workaround to the test. The workaround is
to simply enclose the looops in the test in a block, so that the
DeferStmt will run at the end of that block and before the call to
printMemAllocs.