Adjust sync statements to better handle throws
Consider an example like the following:

  try {
    sync {
      begin someFunction();
      someThrowingFunction();
    }
  } catch {
    ...
  }

Before this PR, that translates into something more or less like this:

  {
    begin someFunction();
    someThrowingFunction(err);
    if err != nil then
      goto handler;
    waitEndCount; // wait for dynamically enclosed begins
  }
  handler:
  if err != nil then
    ...

The problem with this is that if the body of the sync block threw
an error after tasks were created, the sync would not wait for these
tasks. In order to address that issue, this PR wraps the body of each
sync statement in a try/catch block. Additionally, since the sync
statement might 'throw' errors from nested begins, any error 'thrown'
in the body of the sync statement itself needs to be included among
the task errors (otherwise we'd have possibly 2 errors and only one
place to put them - we have to propagate all errors). So, now build.cpp
translates the above example into something like this:

  try {
    try {
      begin someFunction();
      someThrowingFunction(err);
    } catch e:Error {
      chpl_save_task_error(e);
    }
    waitEndCount; // wait for dynamically enclosed begins
  } catch {
    ...
  }