Create a default initializer for first class functions, and a minor bug fix
Now that buildDefaultInitializer can be called outside of buildDefaultFunctions,
call it when we create the class type for a first class function (if we want to
be using default initializers).

Also, update the creation of the super.init() call based on my work to support
inheritance from generic types (I still need a test of this).  I needed to do
it for first class functions because they don't define accessor methods (since
they don't really have fields), so the accessor call to super was failing to
resolve.