Improve error format
Current format looks like this:

uncaught MyError: test error
  e.chpl:5: thrown here
  e.chpl:9: uncaught here

It will take more effort to track functions. In particular,
there is a bit of a design question there because the function
where the error was created/thrown might (often) be deep within
the standard library. How can we show a function name while
avoiding reporting implementation functions? And how can
that work without actually doing a stack trace?

Additionally, I'm inclined to continue to make the error
store where it is thrown vs where it is created because
marking where it is created is connected to initializers &
keeping these efforts separate for now is better.