Fix compilation timeouts/compilation segfaults with --force-initializers
Turns out the timeouts/compiler segfaults were due to an infinite loop after the
handling of initializers on generic types changed, because the "this" argument
wasn't getting labeled appropriately.

Resolve this error by attaching the FLAG_DELAY_GENERIC_EXPANSION to the "this"
arg and "super" field when each is generic.