Adjust error printing
Printing an error should probably result in output similar
to what you get with the uncaught error message. This commit
adjusts the module code to do that by allowing Error instances
to overload the 'proc message()' method that uses 'msg' by default.

In order to make the Error.writeThis apply when Error
is subclassed, I added a workaround to not build writeThis/readThis
for subtypes of Error in buildDefaultFunctions. But this is a
workaround for the problems descibed in issue #7118.