rename ErrorGroup to TaskErrors (#7127)
[reviewed by @mppf]

coforall and forall will eventually emit a flat list of errors from tasks (#7046), so this PR changes the name of that list to TaskErrors instead of ErrorGroup. Note that this PR does not flatten the errors from coforall and forall.

In the future, an ErrorGroup may be made available to users as a way to collect errors separate of the compiler-inserted TaskErrors, either as the parent class of TaskErrors or as a standalone class.